Clean up modified code Rename RentalAddCommand to AddRentalCommand
