+++ src/main/java/seedu/address/logic/Messages.java
@@ -42,8 +42,6 @@ public class Messages {
                 .append(client.getPhone())
                 .append("; Email: ")
                 .append(client.getEmail())
-//                .append("; Address: ")
-//                .append(client.getAddress())
                 .append("; Tags: ");
         client.getTags().forEach(builder::append);
         return builder.toString();
@@ -68,7 +66,7 @@ public class Messages {
                 .append(rentalInformation.getDeposit())
                 .append("; Customer List: ")
                 .append(rentalInformation.getCustomerList());
-//        client.getTags().forEach(builder::append);
+        // client.getTags().forEach(builder::append);
         return builder.toString();
     }
