+++ src/main/java/seedu/address/model/client/Client.java
@@ -8,7 +8,6 @@ import java.util.Objects;
 import java.util.Set;
 
 import seedu.address.commons.util.ToStringBuilder;
-import seedu.address.model.rentalinformation.Address;
 import seedu.address.model.rentalinformation.RentalInformation;
 import seedu.address.model.tag.Tag;
