+++ src/main/java/seedu/address/model/rentalinformation/RentalInformation.java
@@ -39,6 +39,17 @@ public class RentalInformation {
         this.customerList = new CustomerList(customerList);
     }
 
+    /**
+     * Constructs a {@code RentalInformation}.
+     *
+     * @param address A valid {@code Address}.
+     * @param rentalStartDate A valid {@code RentalStartDate}.
+     * @param rentalEndDate A valid {@code RentalEndDate}.
+     * @param rentDueDate A valid {@code RentDueDate}.
+     * @param monthlyRent A valid {@code MonthlyRent}.
+     * @param deposit A valid {@code Deposit}.
+     * @param customerList A valid {@code CustomerList}.
+     */
     public RentalInformation(Address address, RentalDate rentalStartDate, RentalDate rentalEndDate,
                              RentDueDate rentDueDate, MonthlyRent monthlyRent, Deposit deposit,
                              CustomerList customerList) {
@@ -80,13 +91,13 @@ public class RentalInformation {
     }
 
     public String getString() {
-        return "Address: " +  address + "\n" +
-                "Rental start date: " + rentalStartDate + "\n" +
-                "Rental end date: " + rentalEndDate + "\n" +
-                "Rental due date: " + rentDueDate + "\n" +
-                "Monthly rent: " + monthlyRent + "\n" +
-                "Deposit: " + deposit + "\n" +
-                "Customer list: " + customerList + "\n";
+        return "Address: " + address + "\n"
+                + "Rental start date: " + rentalStartDate + "\n"
+                + "Rental end date: " + rentalEndDate + "\n"
+                + "Rental due date: " + rentDueDate + "\n"
+                + "Monthly rent: " + monthlyRent + "\n"
+                + "Deposit: " + deposit + "\n"
+                + "Customer list: " + customerList + "\n";
     }
 
     @Override
