+++ src/main/java/seedu/address/model/util/SampleDataUtil.java
@@ -10,7 +10,6 @@ import seedu.address.model.client.Client;
 import seedu.address.model.client.Email;
 import seedu.address.model.client.Name;
 import seedu.address.model.client.Phone;
-import seedu.address.model.rentalinformation.Address;
 import seedu.address.model.tag.Tag;
 
 /**
@@ -20,23 +19,17 @@ public class SampleDataUtil {
     public static Client[] getSamplePersons() {
         return new Client[] {
             new Client(new Name("Alex Yeoh"), new Phone("87438807"), new Email("alexyeoh@example.com"),
-//                new Address("Blk 30 Geylang Street 29, #06-40"),
-                getTagSet("friends")),
+                    getTagSet("friends")),
             new Client(new Name("Bernice Yu"), new Phone("99272758"), new Email("berniceyu@example.com"),
-//                new Address("Blk 30 Lorong 3 Serangoon Gardens, #07-18"),
-                getTagSet("colleagues", "friends")),
+                    getTagSet("colleagues", "friends")),
             new Client(new Name("Charlotte Oliveiro"), new Phone("93210283"), new Email("charlotte@example.com"),
-//                new Address("Blk 11 Ang Mo Kio Street 74, #11-04"),
-                getTagSet("neighbours")),
+                    getTagSet("neighbours")),
             new Client(new Name("David Li"), new Phone("91031282"), new Email("lidavid@example.com"),
-//                new Address("Blk 436 Serangoon Gardens Street 26, #16-43"),
-                getTagSet("family")),
+                    getTagSet("family")),
             new Client(new Name("Irfan Ibrahim"), new Phone("92492021"), new Email("irfan@example.com"),
-//                new Address("Blk 47 Tampines Street 20, #17-35"),
-                getTagSet("classmates")),
+                    getTagSet("classmates")),
             new Client(new Name("Roy Balakrishnan"), new Phone("92624417"), new Email("royb@example.com"),
-//                new Address("Blk 45 Aljunied Street 85, #11-31"),
-                getTagSet("colleagues"))
+                    getTagSet("colleagues"))
         };
     }
 
@@ -45,6 +38,7 @@ public class SampleDataUtil {
         for (Client sampleClient : getSamplePersons()) {
             sampleAb.addPerson(sampleClient);
         }
+
         return sampleAb;
     }
